requirements.txt
# Blogger News Bot Dependencies
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
feedparser>=6.0.11
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
Pillow>=10.4.0
lxml>=5.2.0
schedule>=1.2.0
pysocks>=1.7.1
